+++ pulsar/src/main/java/com/hazelcast/jet/contrib/pulsar/PulsarSinks.java
@@ -24,7 +24,6 @@ import org.apache.pulsar.client.api.PulsarClient;
 import org.apache.pulsar.client.api.Schema;
 
 import javax.annotation.Nonnull;
-import java.util.Map;
 
 /**
  * Contains factory methods for Pulsar sinks.
@@ -39,7 +38,6 @@ public final class PulsarSinks {
      * a custom Pulsar {@link Sink} for the Pipeline API.
      *
      * @param topic              Pulsar topic name to publish to
-     * @param producerConfig     The configurations are used to create the {@code PulsarProducer}
      * @param connectionSupplier Pulsar client supplier
      * @param extractValueFn     extracts the message value from the emitted items.
      * @param schemaSupplier     Pulsar messaging schema supplier.
@@ -48,29 +46,32 @@ public final class PulsarSinks {
      */
     public static <E, M> PulsarSinkBuilder<E, M> builder(
             @Nonnull String topic,
-            @Nonnull Map<String, Object> producerConfig,
             @Nonnull SupplierEx<PulsarClient> connectionSupplier,
             @Nonnull SupplierEx<Schema<M>> schemaSupplier,
             @Nonnull FunctionEx<? super E, M> extractValueFn
 
     ) {
-        return new PulsarSinkBuilder<>(topic, producerConfig, connectionSupplier, schemaSupplier, extractValueFn);
+        return new PulsarSinkBuilder<>(topic, connectionSupplier, schemaSupplier, extractValueFn);
     }
 
     /**
-     * Convenience for {@link #builder(String, Map, SupplierEx, SupplierEx, FunctionEx)}.
+     * Convenience for {@link #builder(String, SupplierEx, SupplierEx, FunctionEx)}.
      * It creates a basic Pulsar sink that connect the topic.
+     *
+     * @param topic              Pulsar topic name to publish to
+     * @param connectionSupplier Pulsar client supplier
+     * @param schemaSupplier     extracts the message value from the emitted items.
+     * @param extractValueFn     Pulsar messaging schema supplier.
+     * @param <E>                the type of stream items that sink accepts
+     * @param <M>                the type of the message published by {@code PulsarProducer}
      */
     public static <E, M> Sink<E> pulsarSink(
             @Nonnull String topic,
-            @Nonnull String serviceUrl,
-            @Nonnull Map<String, Object> producerConfig,
+            @Nonnull SupplierEx<PulsarClient> connectionSupplier,
             @Nonnull SupplierEx<Schema<M>> schemaSupplier,
             @Nonnull FunctionEx<? super E, M> extractValueFn
     ) {
-        return PulsarSinks.<E, M>builder(topic, producerConfig,
-                () -> PulsarClient.builder().serviceUrl(serviceUrl).build(), schemaSupplier, extractValueFn)
-                .build();
+        return PulsarSinks.<E, M>builder(topic, connectionSupplier, schemaSupplier, extractValueFn).build();
     }
 
 }
